# Conflicts:
#	shop/src/main/java/com/zkxy/shop/ui/goods/AllGoodsActivity.kt
#	shop/src/main/java/com/zkxy/shop/utils/GoodsUtils.kt

+++ shop/src/main/java/com/zkxy/shop/entity/goods/GoodsDetailsEntity.kt
@@ -12,7 +12,7 @@ data class GoodsDetailsEntity(
     val goodsDetails: String?,
     val goodsMoneyPrice: Double?,
     val goodsName: String?,
-    val goodsScorePrice: Int?
+    val goodsScorePrice: Double?,
 ) : Serializable
 
 @Keep
